fix: sécurité, stabilité et performance backend (audit semaines 1-3)
Sécurité:
- JWT secret sorti du code → @Value("${jwt.secret}") / env JWT_SECRET
- DELETE /api/parametres-site et /api/tarifs-circuit-personnalise protégés ADMIN
- Mot de passe tombola → SecureRandom 24 bytes Base64
- Validation magic bytes images (JPEG, PNG, GIF, WebP) dans ImageStorageServiceImpl
- show-sql désactivé, logs Hibernate → WARN

Stabilité:
- @Valid sur @RequestBody CircuitPersonnaliseController et TarifsCircuitPersonnaliseController
- @Transactional ajouté sur 18 services write
- prixTotal et prixParNuit double → BigDecimal (entités, DTOs, services)
- Migration V6: colonnes prix en NUMERIC(10,2)
- Handlers IOException et DataAccessException dans GlobalExceptionHandler
- UtilisateurController: accès direct repository supprimé → passe par le service

Performance:
- @Cacheable sur ZoneService, VilleService, ParametresSiteService + @CacheEvict
- spring-boot-starter-cache + @EnableCaching
- Migration V7: indexes sur statuts et clés étrangères
- @EntityGraph sur DevisRepository, ReservationRepository, ReservationHebergementRepository

+++ src/main/java/com/beninexplo/backend/dto/ReservationHebergementDTO.java
@@ -9,6 +9,7 @@ import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
@@ -50,7 +51,7 @@ public class ReservationHebergementDTO {
     @Positive(message = "Le nombre de personnes doit etre superieur a zero.")
     private int nombrePersonnes;
 
-    private double prixTotal;
+    private BigDecimal prixTotal;
     private String statut;
 
     @Size(max = 1000, message = "Les commentaires ne doivent pas depasser 1000 caracteres.")
@@ -64,7 +65,7 @@ public class ReservationHebergementDTO {
     public ReservationHebergementDTO(Long id, Long hebergementId, String hebergementNom,
                                      String nomClient, String prenomClient, String emailClient,
                                      String telephoneClient, LocalDate dateArrivee, LocalDate dateDepart,
-                                     int nombreNuits, int nombrePersonnes, double prixTotal,
+                                     int nombreNuits, int nombrePersonnes, BigDecimal prixTotal,
                                      String statut, String commentaires, LocalDate dateCreation) {
         this.id = id;
         this.hebergementId = hebergementId;
@@ -121,8 +122,8 @@ public class ReservationHebergementDTO {
     public int getNombrePersonnes() { return nombrePersonnes; }
     public void setNombrePersonnes(int nombrePersonnes) { this.nombrePersonnes = nombrePersonnes; }
 
-    public double getPrixTotal() { return prixTotal; }
-    public void setPrixTotal(double prixTotal) { this.prixTotal = prixTotal; }
+    public BigDecimal getPrixTotal() { return prixTotal; }
+    public void setPrixTotal(BigDecimal prixTotal) { this.prixTotal = prixTotal; }
 
     public String getStatut() { return statut; }
     public void setStatut(String statut) { this.statut = statut; }
